requirements.txt
opencv-python
numpy
matplotlib
pyzmq